Change Token generic type to enum variants
`Token` was generic over type parameter `T`, which was problematic as
there are tokens (most of them actually) with no literal value, so the
type `T` made no sense.

I've switched to an enum for the literal with 2 variants: Number and
String. More can be added later. Decided against using trait for this,
since I can control what kinds of literal there are.