Implement scanning tokens
Since Rust String is encoded in UTF-8, accessing the nth character
(actually Unicode scalar value) is an O(n) operation. For now, I leave
it as is, but this will need to be fixed.

One option might be to use the `source.chars()` iterator directly,
together with the method `next()` and `peek()`. Another is to store
the `chars()` as a Vec inside the scanner. Either way, working with
Unicode requires more care.