Scan identifier and keywords
Since a HashMap (or any type that allocates) cannot be const, I use the
phf crate to create the keywords map.

An alternative is to define a method and match on the text. Another way
is to implement FromStr on TokenType, but this breaks the semantics of
the trait, as some tokens can be converted but should not be.