Add Expr for syntax tree
Add enum Expr containing the expressions in the grammar, each of which
is boxed, as this is a recursive type. Each type of expression is a
struct. I think a trait would be much cleaner; I will try in another
branch later.

I decided to stray from the visitor pattern used in the book, since I
think matching an enum is not that bad. However, this might change.

At first, I attempted to generate the Expr using declarative macros,
but alas that is a terrible way to write code, as you could not even
see the structs you are using. I probably should use proc macros, but
for now, copy and paste is good enough.

(I did pull in a whole crate to derive the `new` constructor, though)