Move printing literal token into its own method
After adding a variant for boolean in the enum LiteralToken, I realize
leaving the printing scattered throughout the codebase is a bad idea.
I would want another more general way without the need for enum, but
this seems to be the best we've got now.

A point of consideration: should a literal with no value be a variant
itself, e.g. `LiteralToken::Null`, or should we use Option for that?