Add ParseError type
Implement Error (Display + Debug) for ParseError. Instead of calling
an `error()` method of Parser that returns a ParseError, construct the
error directly.

One possible drawback of this is it makes the error messages spreaded
among the error types: already, the format strings of the two cases in
the Display impl for ParseError look quite similar. Meanwhile, in the
book, all error reporting is centralized in one place, ensuring the
error format is consistent and leaving chance to improve it further.

Some functions now return Result, which cascade to other functions
calling them: either you handle them properly, or panic on Err. This
is quite close to either try/catch or not doing anything in Java, but
with Exception instead of a Result. Either way, you have to deal with
it anyway, so might as well be explicit.