Add Interpreter to Lox
The Interpreter is fairly close to the Java version, but with one big
difference: using an enum (Value) instead of bare Object, which does not
exist in Rust anyway.

Another change is in the Lox struct. Previously, the `run` method will
create a new Scanner and Parser, so we can pass a mutable reference of
Lox to their methods for error handling. In contrast, the Interpreter
resides in the Lox struct itself, so we cannot pass a reference of the
parent Lox struct to the `interprete` method, since that means multiple
mutable borrows.

Hence, departing from the book, we move the error handling (essentially
setting the two fields of Lox) to the `run` method itself, so we don't
have to pass `self` to Interpreter's method.